Added Fences and Gates to the Pathfinding Scene
Some AI chases the player, and it can be tested that they can follow a
player through gates fine.

A different AI chases a lure enclosed by 4 fences. Right now it waits
patiently once its gotten as close as it has, but I’m going to make it
attack that wall eventually.