Initial Combat, Health, and DayNight Scripts
Press “V” to attack. Will show the ray casts if Debug is toggled on the
player script.

The cube in front of the player can be attacked. Attach a health script
to an object to make it damageable.

Day/Night just Lerps through 4 different colors for day, night,
sunrise, and sunset.